Moved OfflineContentPathResolver to a sub folder
